feat(state): use u8 for field keys and steps; use checked_add/checked_sub on u8; update tests

+++ rust/tests/combinatorics.rs
@@ -3,7 +3,7 @@ use camel_cup::combinatorics::{Permutations, Product};
 #[test]
 fn test_permutations_small() {
     let p = Permutations::new(vec![1, 2, 3]);
-    let mut got: Vec<Vec<i32>> = p.collect();
+    let mut got: Vec<Vec<u8>> = p.collect();
     got.sort(); // canonicalize order for assertion
     let mut expected = vec![
         vec![1, 2, 3],
@@ -20,7 +20,7 @@ fn test_permutations_small() {
 #[test]
 fn test_product_repeat_2() {
     let p = Product::new(vec![1, 2], 2);
-    let mut got: Vec<Vec<i32>> = p.collect();
+    let mut got: Vec<Vec<u8>> = p.collect();
     got.sort();
     let mut expected = vec![vec![1, 1], vec![1, 2], vec![2, 1], vec![2, 2]];
     expected.sort();
